few corrections

+++ defs.h
@@ -1,5 +1,3 @@
-#include "stddef.h"
-
 struct buf;
 struct context;
 struct file;
@@ -25,7 +23,7 @@ void            consoleintr(int(*)(void));
 void            panic(char*) __attribute__((noreturn));
 
 // ether.c
-void ether_send(unsigned char* destMAC,unsigned char* srcMAC,unsigned short type,unsigned char* payload,size_t plen);
+void ether_send(unsigned char* destMAC,unsigned char* srcMAC,unsigned short type,unsigned char* payload,int plen);
 
 // exec.c
 int             exec(char*, char**);
